+++ lib/rucio/db/sqla/migrate_repo/versions/42db2617c364_create_index_on_requests_external_id.py
@@ -1,36 +1,45 @@
-'''
- Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2014-2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' create index on requests.external_id '''
 
- Authors:
- - Vincent Garonne, <vincent.garonne@cern.ch>, 2014-2017
-
-create index on requests.external_id
-
-Revision ID: 42db2617c364
-Revises: 4bab9edd01fc
-Create Date: 2015-03-23 11:56:44.690512
-'''
-
+from alembic import context
 from alembic.op import create_index, drop_index
 
-# revision identifiers, used by Alembic.
-revision = '42db2617c364'  # pylint: disable=invalid-name
-down_revision = '4bab9edd01fc'  # pylint: disable=invalid-name
+
+# Alembic revision identifiers
+revision = '42db2617c364'
+down_revision = '4bab9edd01fc'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('REQUESTS_EXTERNALID_UQ', 'requests', ['external_id'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        create_index('REQUESTS_EXTERNALID_UQ', 'requests', ['external_id'])
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('REQUESTS_EXTERNALID_UQ', 'requests')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        drop_index('REQUESTS_EXTERNALID_UQ', 'requests')
